a lot of reformating changes to make code easier to see. 2 processes merged into one named link_disambiguated_authors. Functions for comparisons/scoring/getting comparison signals are all in disambiguation scoring. Disambiguation_init.py contains all the initialisation logic like blocking and leiden clustering for the first part of the pipeline. util are just non specific algorithms and structures that are being used in the pipeline. Link_ethnicities is a seperate process that links a new relational table that contains the classified ethnicity of each name in the work_author table. Filter ethnicity is a standalone script that works on compressed jsonl data. ethnicity_utils for now contains the classification algorithm used by both the script and the process. Makefile is a temporary mini command just to clone the classfier. Currently working on making the classifier a package so it will be removed when finished
