+++ src/alexandria3k/author_name_disambiguation/makefile
@@ -1,0 +1,4 @@
+CLASSIFIER=https://github.com/name-ethnicity-classifier/name-ethnicity-classifier.git
+DIR=$(HOME)/name-ethnicity-classifier
+clone: 
+	git clone $(CLASSIFIER) $(DIR)
