--- src/alexandria3k/author_name_dissambiguation/celery_tasks.py
@@ -1,3 +1,0 @@
-from celery import Celery
-
-app = Celery('tasks', broker=)
