--- src/alexandria3k/processes/link_merged_author_table.py
@@ -1,431 +1,0 @@
-"""
-Merges authors into disambiguated identities based on shared signals.
-Process takes in populated work_author and author_name_blocks table
-Returns and links merged_authors table (work_author_id , merged_author_id , confidence_score)
-
-Confidence score is calculated by comparing all authors in the same block with eachother,
-Algorithms used to calculate the the confidence score are:
-- Jaccard similarity on co-author sets of each author (how many co-authors they have in common)
-- Jaro Winkler score , comparing the normalized names
-
-Merging authors together is implemented with a UnionFind datastructure in memory and afterwards
-fills the merged_authors table
-"""
-
-from itertools import groupby
-from multiprocessing import Pool
-import time
-from os import cpu_count
-
-import apsw
-
-from rapidfuzz.distance import JaroWinkler
-
-from alexandria3k.common import ensure_table_exists, log_sql, set_fast_writing
-
-from alexandria3k.author_name_dissambiguation.and_utils import (
-    Author,
-    UnionFind,
-    jaccard_similarity,
-    Block_Attr
-)
-
-# from alexandria3k import perf
-from alexandria3k.db_schema import ColumnMeta, TableMeta
-
-table = [
-    TableMeta(
-        "merged_authors",
-        columns=[
-            ColumnMeta("work_author_id"),
-            ColumnMeta("merged_author_id"),
-            ColumnMeta("confidence_score"),
-        ],
-    ),
-]
-
-
-
-def score_affiliations(auth1, auth2, block):
-    """
-    1-3 word n-gram Jaccard similarity of two authors normalized affiliation strings
-    auth = set(author_id, author_name, author_work_id )
-    """
-
-    author_1_affiliations = block.affiliations.get(auth1.id, set())
-    author_2_affiliations = block.affiliations.get(auth2.id, set())
-
-    if not author_1_affiliations and not author_2_affiliations:
-        return None
-
-    return jaccard_similarity(author_1_affiliations, author_2_affiliations)
-
-
-def score_venue(auth1, auth2, block):
-    """
-    Word n-gram Jaccard similarity of the venues two authors published in.
-    """
-
-    author_1_venue = block.venues.get(auth1.id, set())
-    author_2_venue = block.venues.get(auth2.id, set())
-
-    if not author_1_venue and not author_2_venue:
-        return None
-
-    return jaccard_similarity(author_1_venue, author_2_venue)
-
-
-def score_coauthors(auth1, auth2, block, weight=1.5):
-    """Jaccard similarity of two authors' co-author block-key sets"""
-
-    auth1_coauthors = block.co_authors.get(auth1.id, set())
-    auth2_coauthors = block.co_authors.get(auth2.id, set())
-
-    if not auth1_coauthors and not auth2_coauthors:
-        return None
-
-    similarity = score_min_hash(auth1, auth2, block)
-    if similarity is None:
-        similarity = jaccard_similarity(auth1_coauthors, auth2_coauthors)
-
-    return min(1.0, weight * similarity)
-
-
-def score_name_similarity(auth1, auth2, threshold=0.75):
-    """
-    Score the name similarity of 2 authors normalised_name
-    If the name is the same return 1
-    If name1 != name2 find jaroWinkler similarity
-    if jarowinkler < threshold return 0
-    """
-
-    if auth1.name == auth2.name:
-        return 1.0
-    name_similarity = JaroWinkler.similarity(auth1.name, auth2.name)
-
-    return 0 if name_similarity < threshold else name_similarity
-
-def score_min_hash(auth1, auth2, block):
-    """
-    Estimated Jaccard of two authors' co-author sets from their MinHash
-    Returns None when the block was not big enough to build them
-    """
-
-    min_hash1 = block.min_hashes.get(auth1.id)
-    min_hash2 = block.min_hashes.get(auth2.id)
-
-    if min_hash1 is None or min_hash2 is None:
-        return None
-
-    return min_hash1.jaccard(min_hash2)
-
-def check_if_co_authors(auth1, auth2):
-    """
-    Checks if 2 authors are co_authors by comparing if work_id is the same
-    """
-    return auth1.work_id == auth2.work_id
-
-
-def check_communities(auth1, auth2):
-    "Checks if 2 authors are in the same community based on journals"
-    if auth1.community_id is None or auth2.community_id is None:
-        return False
-    return auth1.community_id != auth2.community_id
-
-def check_year_gap(auth1, auth2, block, max_gap=40):
-    """
-    Get year gaps where authors made publications
-    If there is a big gap between them they are probably not the same person
-    """
-
-    year1 = block.publication_years.get(auth1.id)
-    year2 = block.publication_years.get(auth2.id)
-
-    if year1 is None or year2 is None:
-        return None
-
-    gap = abs(year1 - year2)
-    return gap > max_gap
-
-def compare_authors(
-    auth1: Author,
-    auth2: Author,
-    block: Block_Attr,
-    threashold=0.67
-):
-    """This will serve as the scoring function to determine if 2 authors are the same person.
-    The scoring function will be calculated based on a couple of criteria:
-    - Jaccard similarity on co-author sets of each author (how many co-authors they have in common)
-    - Jaro Winkler score , comparing the normalized names
-    - Affiliation/venue overlap
-    - Year gap
-    - Topic overlap using Leiden clustering
-    """
-
-    if check_communities(auth1, auth2):
-        return 0
-    if check_if_co_authors(auth1, auth2):
-        return 0
-    if check_year_gap(auth1, auth2, block):
-        return 0
-
-    name_similarity = score_name_similarity(auth1, auth2)
-    jaccard_affiliations = score_affiliations(auth1, auth2, block)
-    jaccard_coauthors = score_coauthors(auth1, auth2, block)
-    jaccard_venue = score_venue(auth1, auth2, block)
-
-    # calculate confidence score
-    scores = [
-        name_similarity,
-        jaccard_affiliations,
-        jaccard_coauthors,
-        jaccard_venue,
-    ]
-    valid_scores = [s for s in scores if s is not None]
-    avg = sum(valid_scores) / len(valid_scores)
-
-    return avg if avg >= threashold else 0
-
-
-def group_by_signature(authors, block: Block_Attr):
-    """Groups authors by their (name, co-authors, affiliations, year) signature."""
-    groups = {}
-    for author in authors:
-        signature = (
-            author.name,
-            frozenset(block.co_authors.get(author.id, set())),
-            frozenset(block.affiliations.get(author.id, set())),
-            block.publication_years.get(author.id),
-        )
-        if signature not in groups:
-            groups[signature] = []
-        groups[signature].append(author)
-    return groups
-
-
-def process_block(block_key, grouped_authors, database_path, database=None):
-    """
-    Main loop for comparing each author with every other in the block
-    Returns a list of each author entry in the database
-    entry = tuple(work_author_id, merged_id, confidence_score)
-    for comparison details check compare_authors
-    """
-    if database is None:
-        database = apsw.Connection(database_path)
-
-    author_block_list: list[tuple] = []
-    # uses custom Author class for explicitness, set(id , name, work_id)
-    authors = [Author(row[1], row[2], row[3], row[4]) for row in grouped_authors]
-
-    # if block only has one author, skip
-    if len(authors) < 2:
-        author_id = authors[0].id
-        merged_author_id = author_id
-        confidence_score = 1.0
-
-        merged_authors_entry = (author_id, merged_author_id, confidence_score)
-        author_block_list.append(merged_authors_entry)
-        return author_block_list
-
-    # computes everything once per-block for lower query counts
-    block = Block_Attr(block_key, authors)
-    block.load(database)
-
-    # Groups authors by signature, compare only first author of each group
-    groups = group_by_signature(authors, block)
-    representatives = [groups[sign][0] for sign in groups.keys()]
-
-    # Use unionfind dataset for merging
-    union_find = UnionFind(len(groups))
-    scores = [0.0] * len(groups)
-
-    # score = 1 for authors in the same group
-    for i, signature in enumerate(groups.keys()):
-        if len(groups[signature]) > 1:
-            scores[i] = 1.0
-
-    for i, representative_1 in enumerate(representatives):
-        for j, representative_2 in enumerate(representatives):
-
-            if representative_1 == representative_2:
-                continue
-
-            if score := compare_authors(
-                representative_1,
-                representative_2,
-                block
-            ):
-                # dropped check with other merged entries for performance
-                union_find.union(i, j)
-                scores[i] = max(scores[i], score)
-                scores[j] = max(scores[j], score)
-
-    for i, signature in enumerate(groups.keys()):
-        root = union_find.find(i)
-        merged_author_id = representatives[root].id
-        confidence_score = scores[i]
-        for author in groups[signature]:
-            merged_authors_entry = (author.id, merged_author_id, confidence_score)
-            author_block_list.append(merged_authors_entry)
-
-    return author_block_list
-
-
-def process_chunk(chunk, database_path):
-    "Processes a number of blocks in parallel instead of one block at a time"
-
-    database = apsw.Connection(database_path)
-    result = []
-    for block_key, grouped_authors, _ in chunk:
-        result.append(
-            process_block(block_key, grouped_authors, database_path, database)
-        )
-    return result
-
-
-def process_blocks_parallel(block_cursor, database_path, big_block_threashold):
-    """
-    Function is called when create_merged_authors_table "parallelised" flag = True
-    Processes blocks in parallel rather than sequencially
-    Bigger blocks over a threashold are processed in parallel
-    Smaller blocks are divided into chunks and handled in parallel
-    Concurrency occurs with processes not threads
-    Returns a list of block entries in form of a list of tuples
-    """
-
-    start_time = time.perf_counter()
-    big_block_args = []
-    small_block_args = []
-    single_block_results = []
-    query = """SELECT block_key , work_author_id , normalized_name, work_id, community_id
-               FROM author_name_blocks
-               ORDER BY block_key
-            """
-    for block_key, grouped_authors in groupby(
-        block_cursor.execute(query), key=lambda row: row[0]
-    ):
-        grouped_authors = list(grouped_authors)
-        if len(grouped_authors) < 2:
-            work_author_id = grouped_authors[0][1]
-            single_block_results.append([(work_author_id, work_author_id, 1.0)])
-
-        elif len(grouped_authors) > big_block_threashold:
-            big_block_args.append((block_key, grouped_authors, database_path))
-        else:
-            small_block_args.append((block_key, grouped_authors, database_path))
-
-    args_built_time = time.perf_counter()
-    print(f"populated args {args_built_time - start_time:.2f}s")
-    big_block_args.sort(key=lambda args: len(args[1]), reverse=True)
-    small_block_args.sort(key=lambda args: len(args[1]), reverse=True)
-
-    # big blocks get processed one for each process
-    with Pool() as pool:
-        results_big = pool.starmap(process_block, big_block_args, chunksize=1)
-
-    big_blocks_done_time = time.perf_counter()
-    print(f"big blocks {big_blocks_done_time - args_built_time:.2f}s ")
-
-    # split small blocks into a number of workers and process each of them concurrently
-    workers_count = cpu_count()
-    chunks = [small_block_args[j::workers_count] for j in range(workers_count)]
-
-    with Pool(processes=workers_count) as pool:
-        results_small = pool.starmap(
-            process_chunk, [(chunk, database_path) for chunk in chunks]
-        )
-
-    results_small = [row for batch in results_small for row in batch]
-    small_blocks_done_time = time.perf_counter()
-    print(f"small blocks {small_blocks_done_time - big_blocks_done_time:.2f}s ")
-    results = results_big + results_small + single_block_results
-    return results
-
-
-def process_blocks_sequential(block_cursor, database_path, database):
-    """
-    Function is called when create_merged_authors_table "parallelised" flag = False
-    Processes blocks sequencially rather than in parallel
-    Returns a list of block entries in form of a list of tuples
-    """
-    args = []
-    query = """SELECT block_key, work_author_id, normalized_name, work_id, community_id
-               FROM author_name_blocks
-               ORDER BY block_key
-            """
-    for block_key, grouped_authors in groupby(
-        block_cursor.execute(query), key=lambda row: row[0]
-    ):
-        grouped_authors = list(grouped_authors)
-        args.append((block_key, grouped_authors, database_path))
-
-    results = []
-    for block_key, grouped_authors, path in args:
-        results.append(
-            process_block(block_key, grouped_authors, path, database=database)
-        )
-
-    return results
-
-
-def create_merged_authors_table(
-    database_path, big_block_threashold=50, parallelised=True
-):
-    """Creates and links merged_authors table.
-    Takes as input the database path and checks if author_name_blocks table exists
-    """
-
-    database = apsw.Connection(database_path)
-    database.execute(log_sql("DROP TABLE IF EXISTS merged_authors"))
-    database.execute(log_sql(table[0].table_schema()))
-    set_fast_writing(database)
-
-    ensure_table_exists(database, "author_name_blocks")
-    ensure_table_exists(database, "author_affiliations")
-    ensure_table_exists(database, "work_authors")
-    ensure_table_exists(database, "works")
-    # perf.log("merged_authors table created")
-
-    database.execute(log_sql("CREATE INDEX IF NOT EXISTS idx_works_id ON works(id)"))
-    database.execute(
-        log_sql("CREATE INDEX IF NOT EXISTS idx_work_authors_id ON work_authors(id)")
-    )
-    database.execute(log_sql("""
-            CREATE INDEX IF NOT EXISTS idx_author_affiliations_author_id
-            ON author_affiliations(author_id)
-            """))
-    # perf.log("created works/work_authors indexes")
-
-    block_cursor = database.cursor()
-    insert_cursor = database.cursor()
-
-    # perf.log("starting block comparison loop")
-
-    if parallelised:
-        results = process_blocks_parallel(
-            block_cursor, database_path, big_block_threashold
-        )
-    else:
-        results = process_blocks_sequential(block_cursor, database_path, database)
-
-    for rows in results:
-        for row in rows:
-            insert_cursor.execute(
-                "INSERT INTO merged_authors VALUES (?, ?, ?)",
-                row,
-                prepare_flags=apsw.SQLITE_PREPARE_PERSISTENT,
-            )
-
-    # perf.log("finished comparing all blocks")
-
-
-def process(database_path):
-    """Process creates the clusters table for the author name disambiguation pipeline.
-    Input will be the author_names_block table created in
-    /alexandria3k/src/alexandria3k/processes/link_author_blocks.py for less comparisons.
-    For every entry in a block , compares every pair of authors through a scoring function
-    which they will be compared by some criteria mentioned in compare_authors
-    Process will return a table that contains work_author_id , cluster_id , confidence_score
-    where cluster_id is the id of the merged author"""
-
-    create_merged_authors_table(database_path)
